requirements.txt
libphpserialize
click
pycryptodome
requests